Improvements to the problem generation process
* Change how problem is generated. In this version, we cleverly
pupulate some of the tiles instead of just doing  it randomly. In this
way, when we try to solve the board before removing elements, the
board will most likely be solvable. I haven't tested if the initial
board is ALWAYS solvable, though.
* Highest level is reduced to 7, because it turns out that there
should be a minimum number of tiles filled for sudoku to most likely
have unique solution. However, this still doesn't guarantee uniqueness
of the solution.