Change the difficulty leveling system
The difficulty only ranges from 1 to 5 to make coarser
distinction. The main idea of the difficulty is still the same, i.e.,
related to how many tiles empty, but I change how the decision is
made. I set a max and min number of tiles retained, and bin the range
so that we have a lower and upper bounds on how many tiles to remove
in each level. I also added some padding to these bounds, so that
subsequent levels are more distinct in difficulty.